Rewrite event loop and handler to use mailman.
This is a significant patch that removes the entire events package from this library and replaces it with a handler and the mailman listener. It also provides major cleanups to the caching and what properties are stored on a router object. It also makes some efforts to remove a few global variables that should have originally had locks on them.